Revert "init commit"
This reverts commit c6f473c1f6238b7755ea2cc1b591e1eb7029784d.

+++ src/main/java/com/example/musiclibrary/models/BaseEntity.java
@@ -5,11 +5,11 @@ public abstract class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    protected Long id;
-    public Long getId() {
+    protected int id;
+    public int getId() {
         return id;
     }
-    protected void setId(Long id) {
+    protected void setId(int id) {
         this.id = id;
     }
 }
